Revert "cputune: Support cputune for xend driver"
This reverts commit 01692bb167f7ab81213921ba1116d46a4651ef0e.

Quoting the original commit message:

"Not sure if it's the correct way to add cputune xml for xend driver..."

It is not. The defition created that is converted from the internal xend
structures would also be leaked since it isn't used any more.

Revert the commit since it does not make sense to keep the info
internally.